requirements.txt
Django==2.2.6
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
